Solucion de warnings por posible conversion de datos y en Movimientos.cpp porque faltan return false

+++ TRABAJO/Lista_Piezas.h
@@ -24,7 +24,7 @@ public:
 	Color get_color() { return color; }
 	
 	inline void agregar(Pieza* p) { lista.push_back(p); }
-	inline int size() { return lista.size(); }
+	inline int size() { return (int)lista.size(); }
 	
 	void dibuja() { for (auto p : lista) p->dibuja(); }
 	void mueve(int index, Vector2D posicion);
